Second Merge of Scene and UI Work (#3)
* Worked on Chat Manager

* Added Arrays.cs

Added a new script for our username and messages array

* Updated Chat Messages

Made a basic model of what will happen for the chat messages to move

* Finished Basic Chat

Finished all the basics that our chat system will need. Also formatted it to fit on our canvas and make it look realistic.

* Added Blender Files

Added a folder for all our blender work and files. Right now it is just the TV monitor.

* Updated TV Monitor

* Minor Updates

Updated the TVMonitor.blend to render when away from home. Tested TV scale to chat canvas and changed all transform properties so that it looks proper.

* Title Screen with Settings Finished

* Improved UI

* Finished UI Stuff

* Remove dumb files

* Deleted Unnecessary Files

* Final UI Videos. Starting on Text, Starting on Night Scenes

* Finished Up UI Work

* New UI Screens (Maybe Final)

* Polishing Scripts

* Removing Non-needed files

* Removing Conflicts

* Delete Five Nights A Streamer/Assets/Materials/Character Materials directory

* Delete Five Nights A Streamer/Assets/Materials/Character Materials.meta

* Delete Five Nights A Streamer/Assets/Materials directory

* Delete Five Nights A Streamer/Assets/Materials.meta

---------

Co-authored-by: Hogz13 <tc.hart465@k12cpschools.com>