Added Teleportation
Teleportation now works. I added quality things such as an outline on the teleport pad when hovered over and the teleport pad is set to inactive when you teleport on it. I will later add the changing of materials when hovered over and just a general base material for the pads.